Some small cleanup work, added one tentative translation. All done in XMLEditor, with tweaked whitespace handling to avoid unnecessary diffs: • indent: 3 spaces • no extra newlines between children of the same type • added xml:space="preserve" to the lics element to force XMLEditor to NOT rewrap the text
There are still a few mechanical changes coming from XMLEditor: the xhtml namespace declaration is moved from the root element to the element on which it is needed; and one empty line was removed. But now editing with XMLEditor causes mostly real diffs, and not just a lot of whitespace noise.